# backend/internal/api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Container-scoped firewall control plane. Authentication is bearer-token on TCP listeners and peer-credentials on unix sockets.",
        "title": "Firefik API",
        "contact": {},
        "license": {
            "name": "Apache-2.0"
        },
        "version": "0.3"
    },
    "basePath": "/",
    "paths": {
        "/api/agents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/agents on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "List fleet agents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/agents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/agents/{id} on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Get agent detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/agents/{id}/commands": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/agents/{id}/commands on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Send command to an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "command payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/agents/{id}/snapshot": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/agents/{id}/snapshot on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Get agent latest snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/approvals on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "List approvals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/approvals on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Create an approval request",
                "parameters": [
                    {
                        "description": "approval payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/approvals/{id} on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Get an approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/approvals/{id}/approve on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Approve an approval request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "approval decision payload",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/approvals/{id}/reject on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Reject an approval request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "rejection payload",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/enrollment-tokens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/enrollment-tokens on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "List enrollment tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/enrollment-tokens on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Create enrollment token",
                "parameters": [
                    {
                        "description": "agent_id and ttl_seconds",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/templates on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "List control-plane templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/templates on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Publish a control-plane template",
                "parameters": [
                    {
                        "description": "template payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/templates/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/templates/{name} on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Get a control-plane template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "template name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}